IBM RC CycleCloud provider - initial commit
