+++ project.ini
@@ -1,5 +1,5 @@
 [project]
-version = 2.1.1
+version = 3.0.0
 name = lsf
 
 [blobs]
