+++ specs/default/chef/site-cookbooks/lsf/attributes/default.rb
@@ -2,7 +2,7 @@
 default['lsf']['version'] = "10.1"
 default['lsf']['kernel'] = "linux2.6-glibc2.3"
 default['lsf']['arch'] = "x86_64"
-default['lsf']['clustername'] = "azurecc"
+default['lsf']['clustername'] = "cyclecloud"
 default['lsf']['tar_dir'] = "/tmp"
 default['lsf']['lsf_top'] = "/usr/share/lsf"
 default['lsf']['local_etc'] = "/etc/lsf"
