+++ specs/default/chef/site-cookbooks/lsf/recipes/worker.rb
@@ -35,6 +35,13 @@ template "#{node['lsf']['local_etc']}/lsf.conf" do
   )
 end
 
+file "/etc/profile.d/set_env_dir.sh" do
+  content <<-EOH
+export LSF_ENVDIR=#{node['lsf']['local_etc']}
+  EOH
+  mode '644'
+end
+
 defer_block "Defer starting lsf until end of the converge" do
   execute 'lsadmin limstartup' do 
     command "source #{lsf_top}/conf/profile.lsf && LSF_ENVDIR=#{node['lsf']['local_etc']} && lsadmin limstartup -f"
