--- specs/default/chef/site-cookbooks/lsf/templates/conf/azureccprov_config.json.erb
@@ -1,13 +1,0 @@
-{
-    "log_level": "info",
-    "cyclecloud": {
-        "cluster": {
-            "name": "<%= @cycle_clustername %>"
-        },
-        "config": {
-            "username": "<%= @username %>",
-            "password": "",
-            "web_server": "<%= @web_server %>"
-        }
-    }
-}
