--- specs/default/chef/site-cookbooks/lsf/templates/conf/cyclecloud_provider.json.erb
@@ -1,24 +1,0 @@
-{
-    "host_type": "azure_host",
-    "interfaces":
-    [{
-        "name": "getAvailableTemplates",
-        "action": "resource_connector/azurecc/scripts/getAvailableTemplates.sh"
-    },
-    {
-        "name": "getReturnRequests",
-        "action": "resource_connector/azurecc/scripts/getReturnRequests.sh"
-    },
-    {
-        "name": "requestMachines",
-        "action": "resource_connector/azurecc/scripts/requestMachines.sh"
-    },
-    {
-        "name": "requestReturnMachines",
-        "action": "resource_connector/azurecc/scripts/requestReturnMachines.sh"
-    },
-    {
-        "name": "getRequestStatus",
-        "action": "resource_connector/azurecc/scripts/getRequestStatus.sh"
-    }]
-}
