+++ specs/default/chef/site-cookbooks/lsf/templates/conf/cyclecloudprov_config.json.erb
@@ -1,0 +1,8 @@
+{
+    "LogLevel": "INFO",
+    "AZURE_CC_SERVER": "<%= @web_server %>",
+    "AZURE_CC_CLUSTER": "<%= @cycle_clustername %>",
+    "AZURE_CC_USER": "<%= @username %>",
+    "AZURE_CC_PASSWORD": "<%= @password %>",
+    "AZURE_CC_REGION": "<%= @region %>"
+}
