+++ specs/default/chef/site-cookbooks/lsf/templates/conf/hostProviders.json.erb
@@ -1,11 +1,11 @@
 {
   "providers": [
     {
-      "type": "azureProv", 
-      "name": "azurecc", 
-      "scriptPath": "resource_connector/azurecc", 
-      "confPath": "resource_connector/azurecc",
-      "path": "resource_connector/azurecc/provider.json"
+      "type": "cyclecloudProv", 
+      "name": "cyclecloud", 
+      "scriptPath": "resource_connector/cyclecloud", 
+      "confPath": "resource_connector/cyclecloud",
+      "path": "resource_connector/cyclecloud/provider.json"
     }
   ]
 }
