+++ specs/default/chef/site-cookbooks/lsf/templates/conf/lsf.shared.erb
@@ -117,13 +117,13 @@ RESOURCENAME  TYPE    INTERVAL INCREASING  DESCRIPTION 	      # Keywords
 #  egohost    Boolean    ()       ()       (instances from EGO)
 #  awshost    Boolean    ()       ()       (instances from AWS)
 #  softlayerhost Boolean ()       ()       (instances from SoftLayer)
-   azurecchost  Boolean  ()       ()       (instances from Azure CycleCloud)
-   azureccmpi  Boolean   ()       ()       (instances that support MPI placement)
-   azurecclowprio  Boolean ()     ()       (instances that low priority / interruptible from Azure CycleCloud)
+   cyclecloudhost  Boolean  ()       ()       (instances from Azure CycleCloud)
+   cyclecloudmpi  Boolean   ()       ()       (instances that support MPI placement)
+   cyclecloudlowprio  Boolean ()     ()       (instances that low priority / interruptible from Azure CycleCloud)
    ngpus      Numeric    ()       Y       (number of GPUs)
-   nodearray  String     ()       ()       (nodearray from AzureCC)
-   machinetype String    ()       ()       (Azure machine type name for AzureCC)
-   zone       String     ()       ()       (zone/region for AzureCC)
+   nodearray  String     ()       ()       (nodearray from CycleCloud)
+   machinetype String    ()       ()       (Azure machine type name for CycleCloud)
+   zone       String     ()       ()       (zone/region for CycleCloud)
    placementgroup String ()       ()       (id used to note locality of machines)
 #  googlehost Boolean    ()       ()       (instances from Google Cloud)
    rc_account String     ()       ()       (account name for the external hosts)
