--- specs/default/cluster-init/files/host_provider/getRequestStatus.sh
@@ -1,4 +1,0 @@
-#!/bin/bash -e
-scriptDir=$(dirname $0)
-$scriptDir/./invoke_provider.sh create_status $@
-exit $?
