--- specs/default/cluster-init/files/host_provider/getReturnRequests.sh
@@ -1,2 +1,0 @@
-#!/bin/bash -e
-echo '{"message":"This interface is not supported by azurecc provider.", "requests":[]}'
