--- specs/default/cluster-init/files/host_provider/src/version.py
@@ -1,5 +1,0 @@
-__version__ = "1.0.0"
-
-
-def get_version():
-    return __version__
