--- specs/default/cluster-init/files/host_provider/test/test_json_source_helper.py
@@ -1,5 +1,0 @@
-from util import JsonStore
-import sys
-
-if not JsonStore(sys.argv[1], sys.argv[2])._lock():
-    sys.exit(101)
