+++ specs/execute/cluster-init/scripts/00-create-azurecc-profile.sh
@@ -3,13 +3,13 @@
 set -e
 
 lsf_top=$(jetpack config lsf.lsf_top)
-azurecc_profile=$(jetpack config lsf.local_etc /etc/lsf)/azurecc.profile
+cyclecloud_profile=$(jetpack config lsf.local_etc /etc/lsf)/cyclecloud.profile
 
 env_names=$(jetpack config lsf.custom_env_names 0) 2> /dev/null
 
 if [ "$env_names" == "0" ]; then
 	echo no custom environment variables defined.
-	touch $azurecc_profile
+	touch $cyclecloud_profile
 else
 
 	for env_name in $env_names; do
@@ -23,9 +23,9 @@ else
 		
 		ls $lsf_top/conf 1>&2
 		echo "export $env_name=$value" 1>&2
-		echo "export $env_name=$value" >> "$azurecc_profile"
+		echo "export $env_name=$value" >> "$cyclecloud_profile"
 	done
 fi
 
 
-chmod +r $azurecc_profile
+chmod +r $cyclecloud_profile
