+++ specs/master/cluster-init/scripts/020_start_lsf.sh
@@ -1,0 +1,9 @@
+#!/bin/bash
+set -e
+
+lsf_top=$(jetpack config lsf.lsf_top)
+set +e
+source $lsf_top/conf/profile.lsf
+set -e
+
+lsfstartup -f
