+++ templates/lsf.txt
@@ -167,9 +167,9 @@
         PlacementGroupId = ondemandmpi-manual 
 
         [[[configuration]]]
-        lsf.attribute_names = nodearray azureccmpi placementgroup
+        lsf.attribute_names = nodearray cyclecloudmpi placementgroup
         lsf.attributes.nodearray = ondemandmpi
-        lsf.attributes.azureccmpi = true
+        lsf.attributes.cyclecloudmpi = true
         lsf.attributes.placementgroup = ondemandmpi-manual
 
     [[nodearray lowprio]]
@@ -178,9 +178,9 @@
         Interruptible = True
 
         [[[configuration]]]
-        lsf.attribute_names = nodearray azurecclowprio
+        lsf.attribute_names = nodearray cyclecloudlowprio
         lsf.attributes.nodearray = lowprio
-        lsf.attributes.azurecclowprio = true
+        lsf.attributes.cyclecloudlowprio = true
 
     [[nodearray gpu]]
         Priority = 300
@@ -199,9 +199,9 @@
         PlacementGroupId = gpumpi-manual 
 
         [[[configuration]]]
-        lsf.attribute_names = nodearray azureccmpi placementgroup
+        lsf.attribute_names = nodearray cyclecloudmpi placementgroup
         lsf.attributes.nodearray = gpu
-        lsf.attributes.azureccmpi = true
+        lsf.attributes.cyclecloudmpi = true
         lsf.attributes.placementgroup = gpumpi-manual
 
 [parameters About]
@@ -354,7 +354,7 @@ Order = 20
         [[[parameter LsfProjectVersion]]]
         Label = Project Version
         Description = CycleCloud LSF project version found in project.ini
-        DefaultValue = 2.1.1
+        DefaultValue = 3.0.0
 
     [[parameters Master Nodes]]
     Description = "Optional master Fail-over configuration"
